Allow facet attributes to be read.
The binary version of the STL file format allows a 2-byte unsigned
integer that serves as the 'attribute byte count.' This is usually 0,
but some modeling software puts non-zero values in for various
undocumented purposes.

The ASCII version of the STL file format does not have an attribute byte
count field, so this attribute is always unset for ASCII. Round-tripping
between binary and ASCII may therefore lose data.